effect keyword
We add a new keyword to structures and signatures which allows one to
declare algebraic effects as you would with exceptions.

```ocaml
struct
  effect A

  let f () =
    Effect.perform A
```

This is equivalent to writing:

```ocaml
type _ eff += A
```

and is a bit more user-friendly. This also allows local decleration of
effects using `let effect foo in ...`.

Signed-off-by: Ali Caglayan <alizter@gmail.com>